refactor(scripts): extract shared target/module enum conversions for emit runner
`scripts/emit/src/runner.ts` held the string→number parsers (`parseTarget`,
`parseModule`, `inferDefaultModule`) while `cli-transpiler.ts` held the
inverse number→string renderers (`targetToCliArg`, `moduleToCliArg`). Both
files encoded the same ES-year and module-kind tables, so new targets
(e.g. ES2025, Node20) had to be added in two places and could drift.

Move the five functions and their tables into a new
`scripts/emit/src/ts-enums.ts` module; re-import from both callers. Keeps
the existing accept-lenient-but-render-canonical semantics of the parsers
verbatim, and preserves the exact TS6 default-module inference ladder.

No behavior change — `tsc -p scripts/emit/tsconfig.json` clean; emit
runner still discovers/transpiles baselines with the same target/module
resolution.

Tracks DRY_AUDIT_2026-04-21 item: "scripts/emit/src/runner.ts:230,:249,:276
and scripts/emit/src/cli-transpiler.ts:83,:105 repeat target/module tables"
